Updated docker-compose file
Using latest version of all the openCTI images,  so people dont have to edit the file each time there is a new version, not that its that big of a deal but ensuring latest doesnt hurt. Of course pulling the latest compose file is an option but then one has to copy over the AD config. This could of course just be put in an .env file. Anyways, i think using the latest images is something most want.